+++ CHANGELOG.md
@@ -4,6 +4,7 @@ All notable changes to this project will be documented in this file.
 ## [Unreleased]
 ### Changed
 - Customer. Create, Edit. Add "&" in the available signs in the field Customer Account Name.
+- Customer. Create, Edit. Field "City" is implemented. Fields "City", "Line 1" are required.
 
 ## [0.17.0] - 2019-11-21
 ### Changed