+++ CHANGELOG.md
@@ -3,9 +3,14 @@ All notable changes to this project will be documented in this file.
 
 ## [Unreleased]
 ### Changed
+
+## [0.18.0] - 2019-12-13
+### Changed
 - Customers. Create, Edit. Add "&" in the available signs in the field Customer Account Name.
 - Customers. Create, Edit. Field "City" is implemented. Fields "City", "Line 1" are required.
 - Customers. Create, Edit. Placeholders are deleted.
+### Added
+- Module "Customer Comments".
 
 ## [0.17.0] - 2019-11-21
 ### Changed